refactor (dict_factory): :recycle: Move dict_factory to utils.
